ProgramCreator Prototype
ProgramCreator prototype. This commit is able to load and display a program from JSON, this includes the appropriate API calls to load in course data. Additionally, courses can be deleted. To be added is the ability to load in and save a program, edit/add program parts, as well as some finishing touches to the UI.