Saving and Loading
Saving is fully implemented (with the exception of a few lines of CSS to clean up the buttons). Loading is partially implemented, the dialog for the 3 options (Upload, new and Local Storage) in progress - only TypeScript logic left. Current plan is that useProgramCreatorLogic will have the LoadSettings interface, then the LoadProgramDialog will fill this out and pass it back upon the "Upload"/"Create"/"Load" button being pressed.